Yopalito

+++ TallerHelio2/src/co/edu/unbosque/model/CocktailSort.java
@@ -1,14 +1,14 @@
 package co.edu.unbosque.model;
 
 public class CocktailSort {
-	public int[] cocktail_Sort(int arrayNumero[]){
+	public long[] cocktail_Sort(long arrayNumero[]){
 		boolean organizado=true;
 		
 		while(organizado) {
 			organizado = false;
 			for (int x =0;x<=arrayNumero.length-2;x++) {
 				if (arrayNumero[ x ] > arrayNumero[ x + 1 ]) {
-					int temp = arrayNumero[x];
+					long temp = arrayNumero[x];
 					arrayNumero[x] = arrayNumero[x+1];
 					arrayNumero[x+1]=temp;
 					organizado = true;
@@ -20,7 +20,7 @@ public class CocktailSort {
 			organizado = false;
 			for (int x=arrayNumero.length-2;x>=0;x--) {
 				if (arrayNumero[ x ] > arrayNumero[ x + 1 ]) {
-					int temp = arrayNumero[x];
+					long temp = arrayNumero[x];
 					arrayNumero[x] = arrayNumero[x+1];
 					arrayNumero[x+1]=temp;
 					organizado = true;
